First Round of Code Cleanup

+++ src/main/java/kirothebluefox/moblocks/content/ClientForgeRegistration.java
@@ -29,54 +29,54 @@ import net.minecraftforge.fml.common.Mod.EventBusSubscriber.Bus;
 import java.util.ArrayList;
 import java.util.List;
 
-@EventBusSubscriber(modid = MoBlocks.MODID, value = Dist.CLIENT, bus=Bus.FORGE)
+@EventBusSubscriber(modid = MoBlocks.MODID, value = Dist.CLIENT, bus = Bus.FORGE)
 public class ClientForgeRegistration {
-	@SubscribeEvent
-	public static void renderGameOverlayEvent(RenderGameOverlayEvent.Post event) {
-		if (event.getType() == RenderGameOverlayEvent.ElementType.ALL && MoBlocks.config.tooltip_rendering.get()) {
-			PoseStack matrixStack = event.getMatrixStack();
-			HitResult raytraceResult = Minecraft.getInstance().hitResult;
-			if (raytraceResult.getType() == HitResult.Type.BLOCK) {
-				BlockHitResult raytrace = (BlockHitResult) raytraceResult;
-				BlockPos pos = raytrace.getBlockPos();
-				ClientLevel worldIn = Minecraft.getInstance().level;
-				BlockState state = worldIn.getBlockState(pos);
-				Block block = state.getBlock();
-				if (block.getTags().contains(new ResourceLocation(MoBlocks.MODID, "no_gui_container_blocks"))) {
-					Vec3 hit = raytrace.getLocation();
-					ItemStack item = ItemStack.EMPTY;
-					if (block instanceof PotionShelf) {
-						PotionShelf potionShelf = (PotionShelf) block;
-						item = potionShelf.getItemAtHit(worldIn, state, hit, pos);
-					} else if (block instanceof Crate) {
-						Crate crate = (Crate) block;
-						item = crate.getItemAtHit(worldIn, state, hit, pos);
-					} else if (block instanceof Shelf) {
-						Shelf shelf = (Shelf) block;
-						item = shelf.getItemAtHit(worldIn, state, hit, pos);
-					} else if (block instanceof Bookshelf) {
-						Bookshelf bookshelf = (Bookshelf) block;
-						item = bookshelf.getItemAtHit(worldIn, state, hit, pos);
-					}
-					if (!item.isEmpty()) {
-						Font font = Minecraft.getInstance().font;
-						int width = Minecraft.getInstance().getWindow().getGuiScaledWidth(),
-								height = Minecraft.getInstance().getWindow().getGuiScaledHeight(),
-								posx = (int) (width/2),
-								posy = (int) (height/2);
-						//preItemHoverText(item);
-					    List<Component> tooltip = item.getTooltipLines(Minecraft.getInstance().player, Minecraft.getInstance().options.advancedItemTooltips ? TooltipFlag.Default.ADVANCED : TooltipFlag.Default.NORMAL);
-						// MatrixStack, List<ITextComponent>, int, int, int, int, int, FontRenderer
-					    // net.minecraftforge.fml.client.gui.GuiUtils.drawHoveringText(matrixStack, tooltip, posx, posy, width, height, -1, font);
-						List<FormattedCharSequence> tooltipReordering = new ArrayList<>();
-						tooltip.forEach((textcomponent) -> {
-							tooltipReordering.add(FormattedCharSequence.forward(textcomponent.getString(), textcomponent.getStyle()));
-						});
-					    kirothebluefox.moblocks.utils.GuiUtils.renderToolTip(matrixStack, tooltipReordering, posx, posy, font);
-					    //postItemToolTip();
-					}
-				}
-			}
-		}
-	}
+    @SubscribeEvent
+    public static void renderGameOverlayEvent(RenderGameOverlayEvent.Post event) {
+        if (event.getType() == RenderGameOverlayEvent.ElementType.ALL && MoBlocks.config.tooltip_rendering.get()) {
+            PoseStack matrixStack = event.getMatrixStack();
+            HitResult raytraceResult = Minecraft.getInstance().hitResult;
+            if (raytraceResult.getType() == HitResult.Type.BLOCK) {
+                BlockHitResult raytrace = (BlockHitResult) raytraceResult;
+                BlockPos pos = raytrace.getBlockPos();
+                ClientLevel worldIn = Minecraft.getInstance().level;
+                BlockState state = worldIn.getBlockState(pos);
+                Block block = state.getBlock();
+                if (block.getTags().contains(new ResourceLocation(MoBlocks.MODID, "no_gui_container_blocks"))) {
+                    Vec3 hit = raytrace.getLocation();
+                    ItemStack item = ItemStack.EMPTY;
+                    if (block instanceof PotionShelf) {
+                        PotionShelf potionShelf = (PotionShelf) block;
+                        item = potionShelf.getItemAtHit(worldIn, state, hit, pos);
+                    } else if (block instanceof Crate) {
+                        Crate crate = (Crate) block;
+                        item = crate.getItemAtHit(worldIn, state, hit, pos);
+                    } else if (block instanceof Shelf) {
+                        Shelf shelf = (Shelf) block;
+                        item = shelf.getItemAtHit(worldIn, state, hit, pos);
+                    } else if (block instanceof Bookshelf) {
+                        Bookshelf bookshelf = (Bookshelf) block;
+                        item = bookshelf.getItemAtHit(worldIn, state, hit, pos);
+                    }
+                    if (!item.isEmpty()) {
+                        Font font = Minecraft.getInstance().font;
+                        int width = Minecraft.getInstance().getWindow().getGuiScaledWidth(),
+                                height = Minecraft.getInstance().getWindow().getGuiScaledHeight(),
+                                posx = width / 2,
+                                posy = height / 2;
+                        //preItemHoverText(item);
+                        List<Component> tooltip = item.getTooltipLines(Minecraft.getInstance().player, Minecraft.getInstance().options.advancedItemTooltips ? TooltipFlag.Default.ADVANCED : TooltipFlag.Default.NORMAL);
+                        // MatrixStack, List<ITextComponent>, int, int, int, int, int, FontRenderer
+                        // net.minecraftforge.fml.client.gui.GuiUtils.drawHoveringText(matrixStack, tooltip, posx, posy, width, height, -1, font);
+                        List<FormattedCharSequence> tooltipReordering = new ArrayList<>();
+                        tooltip.forEach((textcomponent) -> {
+                            tooltipReordering.add(FormattedCharSequence.forward(textcomponent.getString(), textcomponent.getStyle()));
+                        });
+                        kirothebluefox.moblocks.utils.GuiUtils.renderToolTip(matrixStack, tooltipReordering, posx, posy, font);
+                        //postItemToolTip();
+                    }
+                }
+            }
+        }
+    }
 }
